refactor: Move GitCogniAgent to structured agent directory
Move GitCogniAgent implementation to dedicated directory structure with
proper separation of concerns. Includes organized output directories
for reviews, summaries, and error logs. Update test paths accordingly.
